Move log.py to _internal
