--- docs/api/log.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.log`
-----------------------
-
-.. automodule:: certbot.log
-   :members:
